feat: add a remove_revision() method to SecretRemoveEvent and SecretExpiredEvent (#1624)
In general, charms should treat secret revisions as an internal Juju
detail - they only need to work with the tracked and latest versions,
and don't need to know what numbers those happen to be.

To encourage this behaviour, this PR adds a `remove_revision` method to
the secret-remove and secret-expired events, where charms are expected
to call `secret-remove`. The method needs no arguments because the event
already has the secret and revision in question.

Fixes #1309