feat: add the ability to emit custom events in unit tests (#1589)
In Scenario 7.0, the ability to emit custom events was removed, because
we felt that the proper way to test these was to emit the Juju event
that would trigger them (e.g. relation-changed). However, custom events
are essentially the API for (relation-style) charm libraries, and charms
should not need to know how to populate the relation data to generate
them - only that they will get a specific event, with specific
attributes.

Scenario 6 and earlier could emit custom events, but couldn't provide
any arguments/snapshot for the event, and the user had to provide a
string that identified an attribute on the charm, which doesn't match
with the non-stringly-typed events in Scenario 7. (This functionality is
still present, but requires creating `_Event` objects manually, which
are now private).

Several API choices were considered for emitting custom events, and this
PR differs from the last discussion that was held. The main challenge is
that the ops framework essentially writes a notice of an event (and the
snapshot for that event) into the unit state storage, and then retrieves
notices (and snapshots) and creates the actual event object to pass to
the observing method. This exists to support deferring, but is used even
if the event is not deferred. This requires a `Handle`, and in
particular a path for the `Handle`. In practice, the path is not
determined until the charm object is instantiated (specifically until
the library's `Object` is instantiated), and that is before the
`ctx.run` call, so it is difficult to get matching `Handle` paths,
particularly with only the event object class, which was the planned
API. In addition, the stored observers don't know what type of object
will be emitted, only the event source, which makes something like
`ctx.on.custom(CustomEvent)` difficult to match against the observer
list.

The API added in this PR assumes a charm and library structure like:

```python
# lib/charms/my_lib_charm/v0/lib.py

class MyEvent(ops.EventBase):
    def __init__(self, handle: ops.Handle, arg0: str = "", arg1: int = 0):
        super().__init__(handle)
        self.arg0 = arg0
        self.arg1 = arg1

    def snapshot(self):
        base = super().snapshot()
        base.update({"arg0": self.arg0, "arg1": self.arg1})
        return base

    def restore(self, snapshot: dict[str, typing.Any]):
        super().restore(snapshot)
        self.arg0 = snapshot["arg0"]
        self.arg1 = snapshot["arg1"]

class MyEvents(ops.ObjectEvents):
    foo_changed = ops.EventSource(MyEvent)

class MyConsumer(ops.Object):
    on = MyEvents()

    def __init__(self, charm: ops.CharmBase, relation_name: str):
        super().__init__(charm, relation_name)

# src/charm.py

from my_lib_charm.v0.lib import MyConsumer, MyEvent

class MyCharm(ops.CharmBase):
    def __init__(self, framework: ops.Framework):
        super().__init__(framework)
        self.consumer = MyConsumer(self, "name-of-a-relation")
        framework.observe(self.consumer.on.foo_changed, self._on_event)

    def _on_event(self, event: MyEvent):
        ...
```

Most significantly, the charm must have an attribute that is an `Object`
object, which itself has an attribute that is an `ObjectEvents` object,
and that must be the same type as the one passed to `ctx.run`. I believe
this should cover the standard patterns that libraries use, and if we
need to support others in the future we can extend the API (making it
more complex, but able to accommodate more variation).

This allows emitting a custom event like:

```python
ctx = testing.Context(MyCharm)
state_in = testing.State()
state_out = ctx.run(ctx.on.custom(MyConsumer.on.foo_changed, "foo", arg1=42), state_in)
```

The charm author already needs to know about
`MyConsumer.on.foo_changed`, because they're creating a `MyConsumer`
instance in the charm, and using `on.foo_changed` for the observing.
They also need to know the type of event that they'll get in their
observer method, which implies knowing the arguments/keyword arguments
that are used to create an object of that type. (Note that the Handle
doesn't need to be passed - the framework takes care of that).

(I went through a lot of alternatives to the above - I'm happy to
discuss those or outline some if anyone is interested. I also considered
asking libraries to provide support for testing, but I'd prefer to avoid
that so that use of existing libraries can be tested without those
needing updates, and having to get buy-in for those.).

The implementation looks at the charm object's attributes for an
`Object`, and then looks for an `ObjectEvents` attribute of that, trying
to find one that matches the one provided in the `custom()` call, and
then basically swaps out the unbound event that (e.g.)
`MyConsumer.on.foo_changed` has (missing the framework and handle) for
the bound one that `MyConsumer(...).on.foo_changed` provides. If no
match is found, it falls through to the regular search and will error
out saying that the source can't be found.

This is not dissimilar to the Scenario 6 approach, which used the
(string) event path to navigate through attributes of the charm, but
rather than have to provide the "attribute path" as a string, the charm
author provides the object that is used (essentially) in the observe
call, and Scenario searches for one that matches.

For the event arguments (other than the `Handle`), the implementation
simply bypasses the normal code that figures those out for regular
events. This seems uncontroversial - there isn't really any other way to
get these, unless we had the user provide a snapshot dictionary instead,
which seems less user-friendly (it's more knowing the implementation of
the event than simply its signature).

Some libraries have custom events that take ops objects (particularly
`ops.Relation`) as an argument. The test code cannot pass these to the
`custom()` call, because the framework is needed to get those objects -
the test code should be working with the Scenario options (such as
`testing.Relation`) instead. To facilitate this, Scenario transforms
Scenario objects in the args/kwargs into the (mocked) ops equivalents,
when the framework is available.

There are two "on"s in `ctx.on.custom(MyConsumer.on.foo_changed)`
(assuming that the `ObjectEvents` is called "on", as is usual), which I
don't love, but I prefer this to `ctx.custom(MyConsumer.on.foo_changed)`
or similar.

Some libraries have custom events that inherit from a `HookEvent`
subclass (particularly `RelationEvent`). We advise against this, and do
not support it with the feature in this PR. Most of those events also
need the relation data to be populated as well, so the charm test
already needs to populate a `scenario.Relation` and should trigger the
custom event by running the originating Juju event (such as
`relation_changed`).

Note that if you are testing a *library* then you should not be emitting
custom events in the tests, you should always emit the triggering Juju
ones (having set up the relation data appropriately), and verify that
the custom events are *emitted* correctly.

Drive-bys:
 * Fixed an outdated docstring.
* Removed `_Event.args` and `_Event.kwargs`, which haven't been used for
a while.
* A bunch of tests in `test_context_on` had the wrong name for an object
- mostly this didn't break anything, but it made the tests confusing,
and there was one case where it was a problem.
* To simplify the new tests, added future annotations and switch to
"import ops" for the deferred tests.

Fixes #1421

---------

Co-authored-by: Dave Wilding <tech@dpw.me>