docs: testing explanation (#1635)
The original "Get started with charm testing" how-to:

- "What you will need/learn" (together with the introductory paragraph
that preceeds it) is removed because it fits neither the testing
explanation nor how-tos.
- "Charmcraft profiles" is removed because it's more on bootstrapping a
charm instead of writing tests.
- "Continuous integration" is moved to the explanation page. I wanted to
remove this section altogether because it mostly focuses on how to use
GitHub Actions instead of testing, but after second thought, I think the
examples could give developers a quick start and it has a link to [charm
actions](https://github.com/canonical/charming-actions), so I'm keeping
this section unchanged, only moved to the explanation page.

Testing explanation doc:

- The comment is removed because it only mentions unit/integration while
the doc also mentions interface tests, so it doesn't fit as an
overarching introduction.
- The unit testing section is enriched using content from "Get started
with charm testing", more explanation on unit testing and the test
framework.
- The Interface testing section is enriched using the original
explanation doc on interface tests.
- The integration testing section is enriched using some content from
"Get started with charm testing", mostly the introductory part.
- Stuff on `pytest-operator` is untouched since they would probably be
replaced later by jubilant?

There are two things I am not 100% satisfied but decided to keep anyway:

- The CI examples fit better in a separate how-to guide on "using GitHub
Actions to run charm tests", not in explanations.
- The interface test example probably isn't necessary in the
explanation, but since this example is better than what's in the
how-tos, I want to keep it.

Preview:
https://ops--1635.org.readthedocs.build/en/1635/explanation/testing.html

---------

Co-authored-by: Dave Wilding <tech@dpw.me>