docs: remove tutorial chapters that are covered by the how-to guide (#1511)
The (k8s) tutorial is currently quite long, and we'd like it to be
shorter so that it's easier for charmers to get going.

The plan is that the
[features](https://juju.is/docs/sdk/study-your-application#heading--features)
described in the tutorial should be the chapters. Namely:

* A simple web server with endpoints that can be used, including metrics
 * Integration with postgresql
 * An action to be able to manually interact with postgresql
 * COS integration, including logs and metrics

This PR removes the chapters:
* Exposing the workload version (the existing how-to guide already
covers this, and it's not essential for a charm)
* Storing state (we would generally prefer that charms don't do this, so
shouldn't encourage it, and it'll be covered by a how-to guide before
this PR is ready for review)
* Exposing a port via K8s (too small for a tutorial chapter, and the
existing how-to guide already covers this - and was based mostly on this
chapter, in the knowledge that it'd be replacing it)
* Write Harness tests - the new unit testing (Scenario) chapter is
merged into this. The testing chapters need to be migrated into sections
at the end of each chapter, but that's work for another day and another
person. This just stops teaching new charmers to write Harness tests
* Publishing the charm to Charmhub. Now that this is specifically an ops
tutorial, rather than a broader Juju SDK one, this isn't really part of
our story, and is better placed in the Charmcraft documentation.

There are a few other small tidy-ups, particularly around whitespace and
removing old commented text.

Fixes #1587

---------

Co-authored-by: Dave Wilding <tech@dpw.me>