feat: expose the config metadata in CharmMeta (#1648)
Just like `metadata.yaml` and `actions.yaml` (whether individual files
or in `charmcraft.yaml`), load the `config.yaml` and expose it in
`CharmMeta`.

This adds loading to Harness. Although we're generally not adding
functionality, Harness already loaded the meta to be able to handle
working with config, so it was simpler to just pass it through to
`CharmMeta` than to have it be oddly empty. The one
backwards-incompatible change is that if you had a `config.yaml` that
was missing the required `type` field for an option, Harness would
previously raise a `RuntimeError` on `config_set` and now raises a
`KeyError` when loading the meta. I feel like this is a reasonable
change, and people shouldn't be testing for config that Juju wouldn't
accept.

Scenario also is updated - it was also already setting up `config.yaml`,
but now also passes it through to `CharmMeta`.

Fixes #1493