chore: add changelog and bump versions for release (#1657)
Docs requirements are unchanged.

Observability and docs builds are both expected to fail in this PR due
to the versions of each other that are required by each package not
being released yet. They are temporarily disabled as release
requirements in this PR, and will be restored when bumping the package
versions post-release.

---------

Co-authored-by: Dave Wilding <tech@dpw.me>