ci: post-release versioning, release process update + workflow fix (#1658)
This PR bumps the post-release versions of `ops` and `ops-scenario`, and
updates `HACKING.md` to reflect the process followed in this release. It
also adds a fix for the failing `observability-charm-tests` (needed for
the new versions in this PR), which should fix them for releases too,
and updates the `docs-deps` tox environment that the docs build with the
development version requirements.