docs: fix links to Juju docs (#1681)
Links to https://documentation.ubuntu.com/juju/latest/ now require a
Read the Docs login. This PR updates links to point to
https://documentation.ubuntu.com/juju/3.6/ instead.

Since I'm changing code in the Kubernetes tutorial, I'll also open a
corresponding PR to the
[juju-sdk-tutorial-k8s](https://github.com/canonical/juju-sdk-tutorial-k8s)
repo.

**Note:** This PR doesn't update any links to `juju.is/docs/*`. Those
links already redirect to the 3.6 branch of the docs. They ought to be
updated, but not so urgently. I'll do that separately (#1626).