ci: change prerelease setting used to add latest ops and scenario (#1682)
This PR changes the observability charms tests setup to use
`--prerelease=if-necessary-or-explicit` instead of `--prerelease=allow`. This
is necessary because all the deps are resolved with the prerelease setting, and
we don't want to suddenly prefer prereleases for the other deps. This resolves
the issue with `prometheus-k8s-operator` identified in the update charm pins
PR.