chore: remove code for creating symlinks and handling non-dispatch entry (#1665)
This PR makes the following changes to `_main.py` and the corresponding
unit tests:

- Removes code that creates symlinks for hook executables upon
`install`. See discussion in
https://github.com/canonical/operator/issues/1324.

- Removes the concept of `is_dispatch_aware` within the private
interface of Ops. The code and tests for charm initialisation now assume
that the charm was started via the `dispatch` executable, with
`JUJU_DISPATCH_PATH` set to the requested hook. See
[discussion](https://github.com/canonical/operator/pull/1665#discussion_r2025746968)
and #1677.

- Keeps the ability to detect legacy hook executables and run them.
Although we can't find any examples of charms that need this, we'd
prefer not to break compatibility. See
[discussion](https://github.com/canonical/operator/pull/1665#discussion_r2025746108).

This PR also fixes a Scenario test that started failing (because the
test charm wasn't properly observing an event that was being deferred).
See
[comment](https://github.com/canonical/operator/pull/1665#issuecomment-2773936146).