chore: pyproject.toml changes and blame ignore for ops-scenario ruff formatting (#1694)
This PR adds the config changes that drove the formatting changes made
in #1691. This includes adding a number of `per-file-ignores`, as the
removal of the `tool.ruff.format` table in `testing/pyproject.toml`
means that a number of `ops` linting rules are now being applied to
`ops-scenario` for the first time. These will be fixed in subsequent PRs
(issue: #1693).

This PR also adds the merge commit
(https://github.com/canonical/operator/commit/54f03272439e1e7c450afc5e6b65222f640351fa)
for #1691 to the git blame ignore list (which is why the formatting
changes and config changes are being made in separate PRs).