test: fix overly specific test that fails sometimes with tracing (#1695)
This PR fixes a test that fails intermittently on main. It failed
because it asserted that the last two log lines were specific ones,
while the addition of tracing means that sometimes (but not always),
tracing log lines are emitted afterwards instead. This PR modifies the
test to only assert that both lines are present in the logs instead.

---------

Co-authored-by: Dima Tisnek <dimaqq@gmail.com>