feat: run deferred events with fresh charm instances (#1631)
Create a new charm object for each emission of a deferred event, and for
the event that triggered the ops process.

When using the `testing.Context` as a context manager, the charm that is
made available is the one for the primary event. If there is demand in
the future, we could add a mechanism for accessing the other charm
objects for manipulation prior to running the deferred events, or
inspection after.

This will break charms that are relying on the charm object to be the
same object for deferred and the primary event. However, we have always
told charmers to not do that, and it's risky behaviour because it'll
differ based on whether it's the non-deferred or deferred handler call,
so we're ok with this change.

The other significant change for charms is that, because we need to
commit the framework before finishing up with each charm, the
`pre-commit` and `commit` Lifecycle events are emitted after each
deferred event as well as after the primary event. I'm only aware of two
charms that use these:
* ops-sunbeam, which uses `commit` to put statuses in stored state.
Running more than once should be fine, and the performance impact not
noticeable.
* `statustest` (of @benhoyt's test-charms collection), which seems
almost identical to the sunbeam one. I assume this isn't really used,
and it wouldn't break anyway.

There are three main parts of this implementation:
* Fairly minor adjustments to `_main`, particularly `_Manager`, so that
the charm is passed around and the framework is accessed via the charm,
rather than creating the framework and charm once, and the actual
creation of each charm prior to re-emitting.
* Functionality to "forget" charms. The framework registers event types
on the class, not the instance, so creating multiple charm instances,
even on different `Framework`s, is not normally possible (Scenario gets
around this by wrapping the event source in a new class for each run).
`CharmBase` gains a `_destory_charm` function that undoes all of the
event registration done in `__init__` (for all the dynamic events from
relations, containers, and so forth). I considered using `__del__` for
this, but it's awkward to cleanly remove everything when destruction is
happening, and we don't really want charmers to start creating more than
one charm instance, so this leaves the functionality private.
* Reworking the Scenario "capture events" mechanism, to use a
`Framework` subclass, rather than a context manager. This needed changes
to handle capturing deferred events, but it also fixes a bug.
Previously, the `Framework` *class* was patched, with methods that used
local variables to determine which events would be captured: this meant
that the `capture_events()` context manager was not threadsafe.

Fixes #1174

---------

Co-authored-by: Dima Tisnek <dimaqq@gmail.com>