ci: add zizmor to static check GitHub workflows (#1656)
Changes:
* Adds [zizmor](https://woodruffw.github.io/zizmor/) to pre-commit (for
those using it) and as a workflow.
* Adjust the default permissions for the workflows to none - when more
are needed, each job adds those explicitly (we don't need content:read
because the repo is public).
* Set `persist-credentials` to false for all the checkouts, because we
don't need to save those - we never do follow-up git commands.
* Put the charmcraft GitHub variables into environment variables and use
them from there, which protects against bad expansion.

Many of the workflows are tested as part of this PR. I've tested most of
the others in my fork - the exceptions are TIOBE, publishing (and test
publishing), and updating charm pins. The changes are consistent across
the workflows, so I don't expect any failures with those.