ci: don't run tests within the publish job (#1684)
Assuming that I understood Ben correctly about his earlier conversation
with Tony, we have CI running in the PR context and on the main branch
after each PR is merged.

Thus, we need not run the tests at the start of the publish job.

I'm not sure if this was about all tests or only integration tests.

I've added a step into the release process to wait after the release PR
is merged to make sure tests are indeed passing.