ci: pin workflows by hash (#1721)
This PR updates our uses of actions in our actions and workflows to
specify them by their hash, rather than by tag. This is a good idea
because tags are movable, and so could result in the execution of
malicious code in our workflows if a bad commit is added to any of those
actions.

I believe that dependabot can be configured to open PRs to bump the
commit hash (and version comment) when there are new releases, though
I'm not sure what will happen with our current configuration.