fix: fix link in breakpoint output, remove link from Harness error message (#1726)
This PR addresses a couple of issues:

- In the welcome message that's shown when a breakpoint is triggered,
there's a link to https://juju.is/docs/sdk/debugging. This link is
broken. I think a suitable replacement would be [`juju
debug-code`](https://documentation.ubuntu.com/juju/3.6/reference/juju-cli/list-of-juju-cli-commands/debug-code/).

- Harness raises an error if the charm tries to update an unknown config
option. The error message includes a link to
https://juju.is/docs/sdk/config, which redirects to
[https://ops.readthedocs.io/en/latest/howto/manage-configurations.html](https://ops.readthedocs.io/en/latest/howto/manage-configurations.html).
I was going to replace the link, but since this is Harness, I then
figured that removing the link would be the more maintainable option.