fix: turn on databag access validation in __init__ (#1737)
This PR sets the event context during charm initialisation. The event
name is set to a dummy event name, `'__init__'`. Setting the event name
to a non-falsey value enables relation data access validation during the
charm's `__init__` method.

This means that at runtime, if we don't have permission to access a
databag in `__init__`, we will pre-emptively raise a
`RelationDataAccessError` (a subclass of `ModelError`) like we do in
observers, instead of just raising a plain `ModelError` after the Juju
hook tools run. In Scenario and Harness we will raise a
`RelationDataAccessError` instead of silently passing.

Resolves: #1378