ci: fix reporting to TIOBE after ops-tracing addition (#1755)
A collection of changes to fix reporting to TIOBE after the addition of
the ops-tracing package:

* The TIOBE run needs all the dependencies explicitly installed, so add
the new ones of importlib-metadata and opentelemetry-api. Presumably
when we shift over to using `uv` for dependency management we can clean
this up.
* Add a new tox environment for running unit tests with coverage for
tracing, and split off the reporting part of the existing coverage
environment, and add a new environment for combining and report. This is
very messy, but works for now, and will presumably be able to be cleaned
up when we do the planned repository restructuring.
* Fix one of the issues identified by TIOBE (oddly not by our regular
tooling), where an f-string is used with logging.