fix: correctly remove suffix from event string in ops.testing (#1754)
In `ops.testing`, `scenario.state` contains logic that removes a suffix
string from the event name. With the previous `rsplit` based
implementation, suffix removal would fail if the suffix occurred
multiple times in the string, since it would split the string into more
than two parts.

For example, this would lead to failure when running
`ctx.on.action('my-action')`, as the string `my_action_action` would be
split into `'my', '', ''`, and the code expected the result of splitting
to always be two items (the prefix, and an empty string).

This PR sets `maxsplit=1` for `rsplit` to handle this case.