fix: correctly load an empty Juju config options map (#1778)
There are many forms that "no Juju config options" can take, for
example:

* No `config` section in charmcraft.yaml, and no config.yaml
* An empty `config` section in charmcraft yaml
* An empty config.yaml
* A `config` section in charmcraft yaml with an empty `options`
subsection
* No `config` section in charmcraft.yaml, and an empty `options` map in
config.yaml

For the cases where there *was* an `options` map, but it was empty, we
were failing to load the meta. Add more extensive tests to cover the
different cases, and fix loading in that case.

Fixes #1775

---------

Co-authored-by: Dima Tisnek <dimaqq@gmail.com>