ci: stop smoke testing Charmcraft 2 (#1782)
The smoke tests
[fail](https://github.com/canonical/operator/actions/runs/15235348271/job/42848398516)
when using Charmcraft 2 because of the [setuptools / importlib
changes](https://github.com/canonical/charmcraft/issues/2259), which are
fixed in Charmcraft 3 but not in Charmcraft 2.

Rather than work around this, since we expect charmers to be using
Charmcraft 3 by now, and don't expect charmers to be using very new ops
with old Charmcraft, drop testing against Charmcraft 2.