fix: remote unit data is available in relation-departed (#1364)
During relation-departed the remote unit data should still be available
to the charm, but the relation is no longer included in `relation-list`,
so it's not available by the usual means (but is available when using
the hook tool directly).

When a remote unit name is provided in the Juju context, ensure that
unit is included in the relation units.

Fixes #1109