docs: distribute the ops-scenario README content across the ops docs (#1773)
The existing README for the ops-scenario package (primarily seen on
[PyPI](https://pypi.org/project/ops-scenario/), but also in the release
tarballs, and on GitHub if navigating to the `testing` folder) is a
comprehensive guide (part reference, part how-to) on all of the Scenario
functionality, with a bit of explanation of the design goals thrown in.

A lot of this content is already in the main documentation, but not
100%. The content is distributed as follows:

* Where the text explains how and why Scenario works the way it does,
it's moved to a new explanation doc. This is a separate doc to the
existing testing explanation doc, as that's focused on charm testing
practice in general, and this is more specifically about the design of
Scenario.
* Content that is already well covered by existing how-to and/or
reference docs is removed.
 * Detailed examples are moved to the relevant how-to guide.
* Short examples of how to use a class or explaining its behaviour is
moved to the docstrings/API reference.

For the most part, I've tried to copy chunks of text as they are,
without extensive rewriting - the focus is on *moving* the content
rather than improving it, which can be done in subsequent PRs.
Similarly, there are some how-tos that need a lot of love (like the one
for managing workloads on K8s) and this adds to that without trying to
fix the whole how-to.

Since we no longer have extensive code examples in the README, the tox
environment for validating that code (and the supporting conftest file)
is removed.

A new README is added as a replacement. This is modelled on the [doc
team's model
README](https://docs.google.com/document/d/1Vjx76fB-K39-KH0UZnfOcz3avyGwD5wIS0jA_2MbmbE/edit?tab=t.0#heading=h.rea6xkoe0m4b)
and the [Jubilant
README](https://github.com/canonical/jubilant/blob/main/README.md)
(which I find a very nice modern example in our ecosystem).

[Docs preview](https://ops--1773.org.readthedocs.build/en/1773/),
[README
preview](https://github.com/tonyandrewmeyer/operator/blob/tidy-scenario-readme/testing/README.md)

---------

Co-authored-by: Dave Wilding <tech@dpw.me>