docs: clarify behaviour of Pebble file creation user and group args (#1788)
This PR documents some subtleties of the user and group arguments of the
Pebble file creation methods.

It is an error at runtime to provide only the group to Pebble. This came
up when developing `pathops`, and there are
[tests](https://github.com/canonical/charmtech-charmlibs/blob/main/pathops/tests/integration/pebble/test_container_path.py#L503)
in `pathops` that verify this.

If only the user is specified, the group is set based on the default
user's group. This also came up with `pathops`
([test](https://github.com/canonical/charmtech-charmlibs/blob/main/pathops/tests/integration/pebble/test_container_path.py#L471)).