feat: support for config schema as Python classes (#1741)
A new method is added to `ops.CharmBase`: `load_config`, which is the
custom class equivalent of the `.config` property. The charmer passes a
class to this method and gets back an instance of the class with the
Juju config loaded into it.

The `load_config` method also converts `secret` type config from a `str`
to an `ops.Secret`, since doing anything interesting with the secret
option requires getting the actual secret. This is a 'lazy' Secret, like
`SecretEvent.secret`, that does not have any data from Juju (not even
validation that the secret can be accessed) until it is used.

Two error-handling modes are provided for `load_config`. In the default,
"raise", any exceptions raised creating the class are re-raised and must
be handled by the charm (or left for the charm to go to error state).
The alternative is "blocked", where `load_config` will catch any
`ValueError` (or subclass) and exit the charm with a zero exit status
after setting an appropriate blocked status. This mode is suited for
holistic charms (where handling the specific event is not necessary)
that do not have complex status needs.

The implementation is intended to be duck-typed, in that any class that
accepts the data as keyword arguments and raises ValueError (or a
subclass) in instantiation and/or setting attributes will work. However,
the tests verify four specific cases:

* A generic Python class (no inheritance other than object).
* Standard library dataclasses (I personally believe these are the best
choice for almost all cases).
* Pydantic dataclasses
* Pydantic 2.x BaseModel classes (we'll use these in our documentation,
as they're in common use).

We don't explicitly promise that these exact cases work, but they should
continue to do so while they rely on the current implementation. This
leaves it open to support or not other cases (e.g. a potential Pydantic
3.x, attrs, and so on).

In order to test that these 4 types work, the unit tests add specific
tests if Pydantic is installed, and the tox environment for static and
unit install Pydantic. Ops itself does not require Pydantic (at run time
or for tests).

This PR includes reference docs
([Preview](https://ops--1741.org.readthedocs.build/en/1741/reference/ops.html#ops.CharmBase.load_config))
and updates the configuration how-to
([Preview](https://ops--1741.org.readthedocs.build/en/1741/howto/manage-configuration.html)).
A follow-up PR is planned that would make use of this functionality in
the zero-to-hero tutorial - it will be simplest to do that after #1734
and any follow-up PRs are complete, to avoid conflicts.

More details are available in the
[spec](https://docs.google.com/document/d/1vj2pTvVRXClCrYM7tUjft3ngISSsFdBQD1B-CzOG1Vs/edit?usp=sharing).

---------

Co-authored-by: Ben Hoyt <benhoyt@gmail.com>
Co-authored-by: Ben Hoyt <ben.hoyt@canonical.com>
Co-authored-by: David Wilding <david.wilding@canonical.com>
Co-authored-by: James Garner <james.garner@canonical.com>