feat: support for action parameter schema as Python classes (#1756)
A new method is added to `ops.ActionEvent`: `load_params`, which is the
custom class equivalent of the `.params` property. The charmer passes a
class to this method and gets back an instance of the class with the
Juju action parameters loaded into it.

Two error-handling modes are provided for `load_params`. In the default,
"raise", any exceptions raised creating the class are re-raised and must
be handled by the charm (or left for the charm to go to error state).
The alternative is "fail", where `load_params` will catch any
`ValueError` (or subclass) and exit the charm with a zero exit status
after setting an appropriate failure message. This mode should be
suitable for the majority of actions.

The implementation is intended to be duck-typed, in that any class that
accepts the data as keyword arguments and raises ValueError (or a
subclass) in instantiation and/or setting attributes will work. However,
the tests verify four specific cases:

* A generic Python class (no inheritance other than object).
* Standard library dataclasses (I personally believe these are the best
choice for almost all cases).
* Pydantic dataclasses
* Pydantic 2.x BaseModel classes (we'll use these in our documentation,
as they're in common use).

We don't explicitly promise that these exact cases work, but they should
continue to do so while they rely on the current implementation. This
leaves it open to support or not other cases (e.g. a potential Pydantic
3.x, attrs, and so on).

In order to test that these 4 types work, the unit tests add specific
tests if Pydantic is installed, and the tox environment for static and
unit install Pydantic. Ops itself does not require Pydantic (at run time
or for tests).

This PR includes reference docs
([Preview](https://ops--1756.org.readthedocs.build/en/1756/reference/ops.html#ops.ActionEvent.load_params))
and updates the configuration how-to
([Preview](https://ops--1756.org.readthedocs.build/en/1756/howto/manage-actions.html)).
A follow-up PR is planned that would make use of this functionality in
the zero-to-hero tutorial - it will be simplest to do that after #1734
and any follow-up PRs are complete, to avoid conflicts.

More details are available in the
[spec](https://docs.google.com/document/d/1AiNvr1xfh078ieRczCCc9HSx4VW-dDJdF2p9FS2E_do/edit?usp=sharing).