feat: support for relation data schema as Python classes (#1701)
New `load` and `save` methods are added to `ops.Relation`. `load` takes
a class and returns an instance of that class with the current relation
data loaded into it (the custom class equivalent of the `Relation.data`
attribute). `save` takes an instance of such a class and sends the data
back to Juju. Additional arguments are also able to be passed in if the
class contains more than just the relation data.

In addition, when sending data to Juju, and when loading it into the
class, we automatically serialise/deserialise to/from JSON.

Note that the charmer is responsible for managing multiple instances of
data classes (if both are changed in different ways and then both are
saved, only the last save will be reflected in Juju) and also mixing
with the existing `Relation.data` (which is updated as values change).

Two error-handling modes are provided for `load`. In the default,
"raise", any exceptions raised creating the class are re-raised and must
be handled by the charm (or left for the charm to go to error state).
The alternative is "blocked", where `load` will catch any `ValueError`
(or subclass) and exit the charm with a zero exit status after setting
an appropriate blocked status. The default will be the best choice in
most situations, as if the error is with remote data that most likely
shouldn't change the local unit status, and even when handling local
data status is complicated and a simple "bad data in one relation means
blocked" will only be true in simple situations.

The `errors` argument for `load` primarily exists for consistency with
the `errors` argument of `load_config` and `load_params`. There is no
`errors` argument for `save`: the charm is always expected to handle any
issues.

The implementation is intended to be duck-typed, in that any class that
accepts the data as keyword arguments and raises `ValueError` (or a
subclass) in instantiation and/or setting attributes will work. However,
the tests verify four specific cases:
* A generic Python class (no inheritance other than `object`).
* Standard library dataclasses (I personally believe these are the best
choice for most cases).
* Pydantic dataclasses
* Pydantic 2.x BaseModel classes

We don't explicitly promise that these exact cases work, but they should
continue to do so while they rely on the current implementation. This
leaves it open to support or not other cases (e.g. a potential Pydantic
3.x, `attrs`, and so on).

In order to test that these 4 types work, the unit tests add specific
tests if Pydantic is installed, and the `tox` environment for `static`
and `unit` install Pydantic. Ops itself does *not* require Pydantic (at
run time or for tests).

This PR includes reference docs ([load()
preview](https://ops--1701.org.readthedocs.build/en/1701/reference/ops.html#ops.Relation.load),
[save()
preview](https://ops--1701.org.readthedocs.build/en/1701/reference/ops.html#ops.Relation.save))
and updates the relations how-to
([preview](https://ops--1701.org.readthedocs.build/en/1701/howto/manage-relations.html#set-up-a-relation))
and libraries how-to
([preview](https://ops--1701.org.readthedocs.build/en/1701/howto/manage-libraries.html#write-a-library-that-manages-a-relation-interface)).
A follow-up PR is planned that would make use of this functionality in
the zero-to-hero tutorial - it will be simplest to do that after #1734
and any follow-up PRs are complete, to avoid conflicts.

More details are available in the
[spec](https://docs.google.com/document/d/1G0xfWJXomzzH2VIY1trFq4b5ybMEEjV0w_aReLWyjCs/edit?tab=t.0)
for more details.

---------

Co-authored-by: James Garner <james.garner@canonical.com>
Co-authored-by: Dave Wilding <tech@dpw.me>